Use 16 bits to store LED index
This will allow us to use devices with more than 256 LEDs such as the
LED matrices.

+++ lib/device/DeviceDescription.cpp
@@ -8,7 +8,7 @@ DeviceDescription::DeviceDescription(uint32_t milliamps_supported,
                                      const std::vector<StripDescription> strips)
     : milliamps_supported(milliamps_supported), strips(strips) {}
 
-uint8_t DeviceDescription::GetLedCount() const {
+uint16_t DeviceDescription::GetLedCount() const {
   uint16_t led_count = 0;
   for (const StripDescription& strip : strips) {
     led_count += strip.led_count;
