+++ lib/effect/ColorCycleEffect.cpp
@@ -4,7 +4,7 @@
 
 ColorCycleEffect::ColorCycleEffect() : Effect() {}
 
-CRGB ColorCycleEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
+CRGB ColorCycleEffect::GetRGB(uint16_t led_index, uint32_t time_ms,
                               const StripDescription &strip,
                               RadioPacket *setEffectPacket) const {
   UNUSED(led_index);
