+++ lib/effect/LightningEffect.cpp
@@ -4,7 +4,7 @@
 
 LightningEffect::LightningEffect() : Effect() {}
 
-CRGB LightningEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
+CRGB LightningEffect::GetRGB(uint16_t led_index, uint32_t time_ms,
                              const StripDescription &strip,
                              RadioPacket *setEffectPacket) const {
   UNUSED(led_index);
