+++ lib/effect/RainbowBumpsEffect.cpp
@@ -4,14 +4,14 @@
 
 RainbowBumpsEffect::RainbowBumpsEffect() : Effect() {}
 
-CRGB RainbowBumpsEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
+CRGB RainbowBumpsEffect::GetRGB(uint16_t led_index, uint32_t time_ms,
                                 const StripDescription &strip,
                                 RadioPacket *setEffectPacket) const {
   const uint8_t palette_index =
       setEffectPacket->readPaletteIndexFromSetEffect();
   ColorPalette palette = palettes()[palette_index];
 
-  uint8_t led_count = strip.led_count;
+  uint16_t led_count = strip.led_count;
   if (strip.FlagEnabled(Mirrored)) {
     MirrorIndex(&led_index, &led_count);
   }
