+++ lib/effect/SparkEffect.cpp
@@ -4,7 +4,7 @@
 
 SparkEffect::SparkEffect() : Effect() {}
 
-CRGB SparkEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
+CRGB SparkEffect::GetRGB(uint16_t led_index, uint32_t time_ms,
                          const StripDescription &strip,
                          RadioPacket *setEffectPacket) const {
   const uint8_t pulse_size = brightnesses.size();
@@ -12,7 +12,7 @@ CRGB SparkEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
       setEffectPacket->readPaletteIndexFromSetEffect();
   ColorPalette palette = palettes()[palette_index];
 
-  uint8_t led_count = strip.led_count;
+  uint16_t led_count = strip.led_count;
   if (strip.FlagEnabled(Mirrored)) {
     MirrorIndex(&led_index, &led_count);
   }
